[Improve] Allow swiping to top of scrollview without dismissing
Swiping to the top of the scrollview, and continuing to pan, automatically begins to dismiss the view controller. If the user swipes at velocity, particularly if the scrolling area is small, the velocity pan will continue towards panning the view controller and potentially dismiss it undesirably.
This prevents that and allows a little vertical bounce, unless the scrollview is already at the top